fdfit: move tests into their own files
