+++ lumicks/pylake/fitting/tests/test_model_sim.py
@@ -1,5 +1,6 @@
 import pytest
 import numpy as np
+from lumicks.pylake.fitting.model import Model
 from lumicks.pylake.fitting.models import odijk
 from lumicks.pylake.fitting.parameters import Params, Parameter
 
@@ -33,3 +34,20 @@ def test_simulation_api_wrong_par():
 
     with pytest.raises(KeyError):
         dna([1], {"DNA/Lp": 50.0, "DNA/Lc": 16.0, "DN/St": 1500.0, "kT": 4.11})
+
+
+def test_model_calls():
+    def model_function(x, b, c, d):
+        return b + c * x + d * x * x
+
+    t = np.array([1.0, 2.0, 3.0])
+    model = Model("m", model_function)
+    y_ref = model._raw_call(t, [2.0, 3.0, 4.0])
+
+    np.testing.assert_allclose(model(t, Params(**{"m/a": Parameter(1), "m/b": Parameter(2), "m/c": Parameter(3),
+                                          "m/d": Parameter(4)})), y_ref)
+
+    np.testing.assert_allclose(model(t, Params(**{"m/d": Parameter(4), "m/c": Parameter(3), "m/b": Parameter(2)})), y_ref)
+
+    with pytest.raises(KeyError):
+        np.testing.assert_allclose(model(t, Params(**{"m/a": Parameter(1), "m/b": Parameter(2), "m/d": Parameter(4)})), y_ref)
